Create rules.tsv
Ignore issues as we either have these reasons:
We cannot do this because we have no access to the server sided coding aspect. Nor proxy or the actual web application itself (Apache2)
We use Paypal that does the CSRF stuff for us
Updating Jquery and scrollReveal.js requires to update bootstrap that then also means redoing the site. Which might happen depending on our front-end developer :)